requirements.txt
pandas
google-cloud-bigquery
requests
moz-sql-parser
requests
pyarrow
pandas_gbq